Title: 1까지 나누기, Time: 121ms, Memory: 10MB, Status: Passed - Codetree

+++ 250531/1까지 나누기/divide-by-1.java
@@ -11,8 +11,8 @@ public class Main {
         for (int i = 1; ; i++) {
 
             cnt++;
-            
-            if (a / (double) i <= 1) {
+
+            if (a / i <= 1) {
                 break;
             }
